Split tests to their own files.
It required making a bunch of internal types public though.
