+++ src/lib.rs
@@ -2,3 +2,8 @@
 mod lexer;
 pub mod pp;
 mod token;
+
+#[cfg(test)]
+mod lexer_tests;
+#[cfg(test)]
+mod pp_tests;
